+++ src/Extension.php
@@ -3,8 +3,9 @@
 /**
  * Title: WordPress pay extension Gravity Forms extension
  * Description:
- * Copyright: Copyright (c) 2005 - 2015
+ * Copyright: Copyright (c) 2005 - 2016
  * Company: Pronamic
+ *
  * @author Remco Tolsma
  * @version 1.3.0
  * @since 1.0.0
@@ -222,7 +223,11 @@ class Pronamic_WP_Pay_Extensions_GravityForms_Extension {
 						break;
 				}
 
-				Pronamic_WP_Pay_Extensions_GravityForms_GravityForms::update_entry( $lead );
+				Pronamic_WP_Pay_Extensions_GravityForms_GravityForms::update_entry_property(
+					$lead['id'],
+					Pronamic_WP_Pay_Extensions_GravityForms_LeadProperties::PAYMENT_STATUS,
+					$lead[ Pronamic_WP_Pay_Extensions_GravityForms_LeadProperties::PAYMENT_STATUS ]
+				);
 
 				if ( $url && $can_redirect ) {
 					wp_redirect( $url );