+++ src/PaymentFormPostType.php
@@ -3,8 +3,9 @@
 /**
  * Title: WordPress payment form post type
  * Description:
- * Copyright: Copyright (c) 2005 - 2015
+ * Copyright: Copyright (c) 2005 - 2016
  * Company: Pronamic
+ *
  * @author Remco Tolsma
  * @version 1.1.0
  */