+++ src/PaymentMethodSelector.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * Title: WordPress pay extension Gravity Forms issuer drop down
+ * Title: WordPress pay extension Gravity Forms payment method selector
  * Description:
  * Copyright: Copyright (c) 2005 - 2016
  * Company: Pronamic
@@ -9,13 +9,13 @@
  * @author Remco Tolsma
  * @version 1.0.0
  */
-class Pronamic_WP_Pay_Extensions_GravityForms_IssuerDropDown {
+class Pronamic_WP_Pay_Extensions_GravityForms_PaymentMethodSelector {
 	/**
 	 * Type
 	 *
 	 * @var string
 	 */
-	const TYPE = 'ideal_issuer_drop_down';
+	const TYPE = 'pronamic_pay_payment_method_selector';
 
 	//////////////////////////////////////////////////